requirements.txt
django-environ
dj-database-url
django-heroku
psycopg2
whitenoise
gunicorn
Pillow